FEATURE: Icon in typescript
